hal_sd: Implement flush
To speed up writing data to the SD card, hal_sd
does not write until a sector is evicted from the
cache or hal_flush is called.
hal_flush is called when a file is closed and at
the end of a call to fat16_write. Only sectors
which have been modified are written back to the
SD card.

Signed-off-by: Francois Berder <fberder@outlook.fr>